filtering works
