+++ src/App.css
@@ -60,7 +60,6 @@
   background-color: #A30000;
   color: white;
   text-align: center;
-  padding: 10px;
 }
 
 nav a{
@@ -68,6 +67,36 @@ nav a{
   color: white;
   text-align: left;
 }
+.filter-nav-section{
+  background-color: #6B0000;
+  display: flex;
+  width: 100%;
+  padding: 10px 10px;
+  
+}
+
+.search-bar-nav{
+  float: right;
+  padding-left: 200px;
+}
+.search-bar-nav button{
+  margin-left: 30px;
+}
+.form-css input {
+  padding: 5px 15px;
+  border-radius: 10px;
+  background-color: #ffffff;
+  color: rgb(126, 126, 126);
+  box-shadow: 0 2px 3px rgba(0, 0, 0, 0.1);
+}
+
+.filter-and-search{
+display: flex;
+  margin: 0;
+  padding: 0;
+  width: 100%;
+}
+
 
 main{
   color: white;
@@ -109,10 +138,11 @@ main{
 }
 
 .custom-select {
-  padding: 8px 12px;
+  padding: 6px 10px;
   border: 1px solid #ced4da;
-  border-radius: 4px;
-  background-color: white;
+  border-radius: 5px;
+  background-color: #6B0000;
+  color: white;
   box-shadow: 0 2px 3px rgba(0, 0, 0, 0.1);
   font-size: 16px;
   margin-right: 10px;
@@ -125,8 +155,8 @@ main{
 }
 
 .custom-button {
-  background-color: #ff4d4d;
-  padding: 10px 15px;
+  background-color: #5B5B5B;
+  padding: 3px 10px;
   color: white;
   border: none;
   border-radius: 4px;
