+++ src/Pages/Categories.js
@@ -7,11 +7,13 @@ import brake from './../Components/Images/category_brake.png';
 import suspension from './../Components/Images/category_suspension.png'; 
 import SiteHeader from '../Components/SiteHeader';
 import SiteFooter from '../Components/SiteFooter';
-
+import { CarDetails } from '../Components/Constants';
 
 const CategoryList = () => {
   const [categories, setCategories] = useState([]);
   const navigate = useNavigate();
+
+  
 
   const handleProductsList = (categoryId) => {
     navigate(`/categories/${categoryId}/products`);
